uucore: extendedbigdecimal: Add from_number_exp2 to compute num * 2**exp
Move the code from num_parser, as we have another usecase in od.
