+++ src/uucore/src/lib/features/extendedbigdecimal.rs
@@ -2,7 +2,7 @@
 //
 // For the full copyright and license information, please view the LICENSE
 // file that was distributed with this source code.
-// spell-checker:ignore bigdecimal extendedbigdecimal biguint
+// spell-checker:ignore bigdecimal extendedbigdecimal biguint prec ilog muls
 //! An arbitrary precision float that can also represent infinity, NaN, etc.
 //!
 //! The finite values are stored as [`BigDecimal`] instances. Because
@@ -21,11 +21,11 @@
 //! assert_eq!(summand1 + summand2, ExtendedBigDecimal::Infinity);
 //! ```
 use std::cmp::Ordering;
+use std::num::NonZeroU64;
 use std::ops::Add;
 use std::ops::Neg;
 
-use bigdecimal::BigDecimal;
-use bigdecimal::num_bigint::BigUint;
+use bigdecimal::{BigDecimal, Context, num_bigint::BigUint};
 use num_traits::FromPrimitive;
 use num_traits::Signed;
 use num_traits::Zero;
@@ -113,6 +113,71 @@ impl From<u32> for ExtendedBigDecimal {
     }
 }
 
+/// Compute bd**exp using exponentiation by squaring algorithm, while maintaining the
+/// precision specified in ctx (the number of digits would otherwise explode).
+///
+/// Algorithm comes from <https://en.wikipedia.org/wiki/Exponentiation_by_squaring>
+///
+/// TODO: Still pending discussion in <https://github.com/akubera/bigdecimal-rs/issues/147>,
+/// we do lose a little bit of precision, and the last digits may not be correct.
+/// Note: This has been copied from the latest revision in <https://github.com/akubera/bigdecimal-rs/pull/148>,
+/// so it's using minimum Rust version of `bigdecimal-rs`.
+pub fn pow_with_context(bd: &BigDecimal, exp: i64, ctx: &Context) -> BigDecimal {
+    if exp == 0 {
+        return 1.into();
+    }
+
+    // When performing a multiplication between 2 numbers, we may lose up to 2 digits
+    // of precision.
+    // "Proof": https://github.com/akubera/bigdecimal-rs/issues/147#issuecomment-2793431202
+    const MARGIN_PER_MUL: u64 = 2;
+    // When doing many multiplication, we still introduce additional errors, add 1 more digit
+    // per 10 multiplications.
+    const MUL_PER_MARGIN_EXTRA: u64 = 10;
+
+    fn trim_precision(bd: BigDecimal, ctx: &Context, margin: u64) -> BigDecimal {
+        let prec = ctx.precision().get() + margin;
+        if bd.digits() > prec {
+            bd.with_precision_round(NonZeroU64::new(prec).unwrap(), ctx.rounding_mode())
+        } else {
+            bd
+        }
+    }
+
+    // Count the number of multiplications we're going to perform, one per "1" binary digit
+    // in exp, and the number of times we can divide exp by 2.
+    let mut n = exp.unsigned_abs();
+    // Note: 63 - n.leading_zeros() == n.ilog2, but that's only available in recent Rust versions.
+    let muls = (n.count_ones() + (63 - n.leading_zeros()) - 1) as u64;
+    // Note: div_ceil would be nice to use here, but only available in recent Rust versions.
+    // (see note above about minimum Rust version in use)
+    let margin_extra = (muls + MUL_PER_MARGIN_EXTRA / 2) / MUL_PER_MARGIN_EXTRA;
+    let mut margin = margin_extra + MARGIN_PER_MUL * muls;
+
+    let mut bd_y: BigDecimal = 1.into();
+    let mut bd_x = if exp >= 0 {
+        bd.clone()
+    } else {
+        bd.inverse_with_context(&ctx.with_precision(
+            NonZeroU64::new(ctx.precision().get() + margin + MARGIN_PER_MUL).unwrap(),
+        ))
+    };
+
+    while n > 1 {
+        if n % 2 == 1 {
+            bd_y = trim_precision(&bd_x * bd_y, ctx, margin);
+            margin -= MARGIN_PER_MUL;
+            n -= 1;
+        }
+        bd_x = trim_precision(bd_x.square(), ctx, margin);
+        margin -= MARGIN_PER_MUL;
+        n /= 2;
+    }
+    debug_assert_eq!(margin, margin_extra);
+
+    trim_precision(bd_x * bd_y, ctx, 0)
+}
+
 impl ExtendedBigDecimal {
     pub fn zero() -> Self {
         Self::BigDecimal(0.into())
@@ -134,6 +199,20 @@ impl ExtendedBigDecimal {
             }
             _ => None,
         }
+    }
+
+    /// Create an `ExtendedBigDecimal` whose value is `number * 2**exponent`. Note that some
+    /// precision may be lost for large exponent, to prevent the representation from blowing up.
+    pub fn from_number_exp2<T: std::ops::Mul<BigDecimal, Output = BigDecimal>>(
+        number: T,
+        exponent: i64,
+    ) -> Self {
+        let base: BigDecimal = 2.into();
+        // Note: We cannot overflow/underflow BigDecimal here, as we will not be able to reach the
+        // maximum/minimum scale (i64 range).
+        let pow2 = pow_with_context(&base, exponent, &Context::default());
+
+        ExtendedBigDecimal::BigDecimal(number * pow2)
     }
 }
